+++ src/sdk/sdk.ts
@@ -61,9 +61,9 @@ export class SDKConfiguration {
     serverDefaults: any;
     language = "typescript";
     openapiDocVersion = "latest";
-    sdkVersion = "0.4.2";
-    genVersion = "2.258.2";
-    userAgent = "speakeasy-sdk/typescript 0.4.2 2.258.2 latest @lukehagar/discoursejs";
+    sdkVersion = "0.4.3";
+    genVersion = "2.272.4";
+    userAgent = "speakeasy-sdk/typescript 0.4.3 2.272.4 latest @lukehagar/discoursejs";
     retryConfig?: utils.RetryConfig;
     public constructor(init?: Partial<SDKConfiguration>) {
         Object.assign(this, init);